refact: move default implementation to interfaces
