+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/searchbox/SearchBoxCompose.kt
@@ -1,11 +1,13 @@
 package com.algolia.instantsearch.compose.searchbox
 
+import androidx.compose.runtime.Stable
 import com.algolia.instantsearch.compose.searchbox.internal.SearchBoxComposeImpl
 import com.algolia.instantsearch.core.searchbox.SearchBoxView
 
 /**
  * Search Box view for compose.
  */
+@Stable
 public interface SearchBoxCompose : SearchBoxView {
 
     /**
@@ -15,8 +17,19 @@ public interface SearchBoxCompose : SearchBoxView {
 
     /**
      * the callback to be triggered on text update
+     *
+     * @param query received query input
+     * @param isSubmit true if the value is a submit, otherwise false
      */
-    public fun onValueChange(query: String, isSubmit: Boolean)
+    public fun onValueChange(query: String, isSubmit: Boolean) {
+        val onQuery = if (isSubmit) onQuerySubmitted else onQueryChanged
+        onQuery?.invoke(query)
+    }
+
+    override fun setText(text: String?, submitQuery: Boolean) {
+        query = text ?: ""
+        if (submitQuery) onQuerySubmitted?.invoke(text)
+    }
 }
 
 /**
